requirements.txt
SQLAlchemy==1.2.5
lxml==4.1.1
requests==2.18.4
python_dateutil==2.7.2
typing==3.6.4
